KS-3212: RepositoryException when stopping tomcat PLF 3.5 M2

+++ eXoApplication/wiki/service/src/main/java/org/exoplatform/wiki/service/impl/SessionDestroyedListener.java
@@ -52,10 +52,13 @@ public class SessionDestroyedListener extends Listener<PortalContainer, HttpSess
     if (LOG.isTraceEnabled()) {
       LOG.trace("Removed the key: " + sessionId);
     }
-
-    WikiService wikiService = (WikiService) PortalContainer.getComponent(WikiService.class);
-    RequestLifeCycle.begin(PortalContainer.getInstance());
-    wikiService.deleteDraftNewPage(sessionId);
-    RequestLifeCycle.end();
+    
+    PortalContainer container = event.getSource();
+    if (container.isStarted()) {
+      WikiService wikiService = (WikiService) container.getComponentInstanceOfType(WikiService.class);
+      RequestLifeCycle.begin(PortalContainer.getInstance());
+      wikiService.deleteDraftNewPage(sessionId);
+      RequestLifeCycle.end();
+    }
   }
 }
